fix: Simplified the error directory
