--- src/errors/tekton_error.rs
@@ -1,24 +1,0 @@
-use core::fmt;
-
-#[derive(Debug, Clone)]
-pub enum TektonError {
-    Reason(String),
-}
-
-impl fmt::Display for TektonError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let string = match self {
-            TektonError::Reason(r) => r.clone(),
-        };
-
-        write!(f, "{}", string)
-    }
-}
-
-#[test]
-fn test_tekton_error_enum() {
-    let reason = "test error".to_string();
-    let err = TektonError::Reason(reason.clone());
-    let string_err = err.to_string();
-    assert_eq!(string_err, reason);
-}
